feat(stage1): add shared fastify server
