+++ Docs/Implementation/implementation-stages.md
@@ -90,7 +90,7 @@ gantt
 
 - [x] Implement base entity classes and domain models
 - [x] Create repository pattern interfaces and PostgreSQL implementations
-- [ ] Set up Fastify applications for CP and DP
+- [x] Set up Fastify applications for CP and DP
 - [ ] Implement dependency injection container
 - [ ] Create event bus for internal communication
 - [ ] Set up database migrations with Prisma
