+++ Docs/Implementation/implementation-summary.md
@@ -73,6 +73,7 @@ This project follows a comprehensive documentation architecture designed for bot
 - ✅ **CI/CD & Dev Environment**: GitHub Actions pipeline and Docker setup with PostgreSQL and Redis
 - ✅ **Core Domain Models**: Base entity classes with Participant and Asset definitions
 - ✅ **Repository Layer**: Interfaces and PostgreSQL implementations for core entities
+- ✅ **Fastify Applications**: Shared Fastify setup for Control and Data Planes with health endpoints
 
 ## Implementation Roadmap
