+++ packages/control-plane/src/index.ts
@@ -1,31 +1,15 @@
-import fastify from 'fastify';
-import { ConnectorError } from '../../core/src/errors';
+import { createServer } from '@connector/core';
 
-const server = fastify();
+/**
+ * Bootstraps and starts the Control Plane Fastify server.
+ */
+export async function start(): Promise<void> {
+  const server = createServer();
+  const port = Number(process.env.CP_PORT) || 3000;
 
-server.get('/health', async () => {
-  return { status: 'ok' };
-});
-
-server.setErrorHandler((error, request, reply) => {
-  if (error instanceof ConnectorError) {
-    const connectorError = error as ConnectorError;
-    reply.status(connectorError.statusCode).send({
-      error: connectorError.errorCode,
-      message: connectorError.message,
-    });
-  } else {
-    reply.status(500).send({
-      error: 'INTERNAL_SERVER_ERROR',
-      message: 'Internal server error',
-    });
-  }
-});
-
-export async function start() {
   try {
-    await server.listen({ port: 3000, host: '0.0.0.0' });
-    console.info('Control Plane server running at http://localhost:3000');
+    await server.listen({ port, host: '0.0.0.0' });
+    console.info(`Control Plane server running at http://localhost:${port}`);
   } catch (err) {
     console.error('Error starting Control Plane server:', err);
     process.exit(1);
