+++ packages/core/src/utils/index.ts
@@ -14,3 +14,5 @@ export function isValidUrl(url: string): boolean {
     return false;
   }
 }
+
+export { createServer } from './server';
