+++ packages/data-plane/src/index.ts
@@ -1,30 +1,15 @@
-import fastify from 'fastify';
-import { ConnectorError } from '../../core/src/errors';
+import { createServer } from '@connector/core';
 
-const server = fastify();
+/**
+ * Bootstraps and starts the Data Plane Fastify server.
+ */
+export async function start(): Promise<void> {
+  const server = createServer();
+  const port = Number(process.env.DP_PORT) || 3001;
 
-server.get('/health', async () => {
-  return { status: 'ok' };
-});
-
-server.setErrorHandler((error, request, reply) => {
-  if (error instanceof ConnectorError) {
-    reply.status(error.statusCode).send({
-      error: error.errorCode,
-      message: error.message,
-    });
-  } else {
-    reply.status(500).send({
-      error: 'INTERNAL_SERVER_ERROR',
-      message: 'Internal server error',
-    });
-  }
-});
-
-export async function start() {
   try {
-    await server.listen({ port: 3001, host: '0.0.0.0' });
-    console.info('Data Plane server running at http://localhost:3001');
+    await server.listen({ port, host: '0.0.0.0' });
+    console.info(`Data Plane server running at http://localhost:${port}`);
   } catch (err) {
     console.error('Error starting Data Plane server:', err);
     process.exit(1);
